refactor: IPNE 描画のプレイヤー層を drawPlayer へ分離
- プレイヤー本体・オーラ・シールド・残像・武器光跡・衝撃波・パーティクルを逐語移植
- afterImageManagerRef の記録/描画の副作用と優先度分岐を保存
- 特性化テストで描画コマンド列の不変を確認
