+++ src/features/ipne/presentation/screens/render/renderGameFrame.ts
@@ -27,23 +27,14 @@ import {
   WALL_SPRITE,
   getStageFloorSprite,
   getStageWallSprite,
-  getPlayerSpriteSheet,
-  WARRIOR_ATTACK_SPRITE_SHEETS,
-  THIEF_ATTACK_SPRITE_SHEETS,
-  WARRIOR_DAMAGE_SPRITES,
-  THIEF_DAMAGE_SPRITES,
-  WARRIOR_IDLE_SPRITE_SHEETS,
-  THIEF_IDLE_SPRITE_SHEETS,
 } from '../../sprites';
-import { drawPlayerAura } from '../../effects/aura';
-import { drawWeaponTrail, getWeaponTier, WeaponTier, drawShockwave } from '../../effects/weaponEffect';
-import { drawShieldGlow, drawAfterImage, drawSpinParticles, drawHealParticles } from '../../effects/stageVisual';
 import { getStageIntroPhase, getStageIntroAlpha, getStageIntroTextAlpha, getGameOverTransitionAlpha } from '../../effects/screenTransition';
 import type { RenderContext, FrameContext } from './renderContext';
 import type { Position, Viewport } from '../../../index';
 import { drawWorld } from './drawWorld';
 import { drawEnemies } from './drawEnemies';
 import { combatEffects } from './combatEffects';
+import { drawPlayer } from './drawPlayer';
 
 /**
  * ゲームフレームを描画する
@@ -65,17 +56,10 @@ export function renderGameFrame(rc: RenderContext): void {
     debugState,
     isDying,
     currentStage,
-    rewardEffects,
-    spriteRenderer,
-    movementStateRef,
     effectManagerRef,
-    deathEffectRef,
     bossWarningRef,
-    afterImageManagerRef,
     stageStartTimeRef,
     dyingStartTimeRef,
-    playerAttackUntilRef,
-    playerDamageUntilRef,
     comboStateRef,
   } = rc;
 
@@ -183,115 +167,8 @@ export function renderGameFrame(rc: RenderContext): void {
   // 戦闘エフェクト処理（攻撃/被弾トリガー・外部キュー・エフェクト更新描画・フローティングテキスト）
   combatEffects(frame);
 
-  // プレイヤー描画（T-02.4: スプライト描画）
-  const deathEff = deathEffectRef.current;
-  const playerDrawSize = SPRITE_SIZES.base * spriteScale;
-
-  if (isDying && deathEff.isActive()) {
-    // 死亡アニメーション中
-    const playerColors = player.playerClass === 'warrior'
-      ? ['#667eea', '#5a67d8', '#4c51bf', '#ffffff']
-      : ['#a78bfa', '#8b5cf6', '#7c3aed', '#ffffff'];
-
-    deathEff.update(now, playerScreen.x, playerScreen.y, playerColors);
-
-    // フェーズに応じてプレイヤースプライトを表示/非表示
-    if (deathEff.isPlayerVisible(now)) {
-      const playerSheet = getPlayerSpriteSheet(
-        player.playerClass as 'warrior' | 'thief',
-        player.direction as 'down' | 'up' | 'left' | 'right'
-      );
-      const playerDrawX = playerScreen.x - playerDrawSize / 2;
-      const playerDrawY = playerScreen.y - playerDrawSize / 2;
-
-      // 待機フレームで描画
-      spriteRenderer.drawSprite(ctx, playerSheet.sprites[0], playerDrawX, playerDrawY, spriteScale);
-    }
-
-    // 死亡エフェクト描画（赤変色オーバーレイ + パーティクル分解）
-    deathEff.draw(ctx, now, playerScreen.x, playerScreen.y, playerDrawSize);
-  } else {
-    // 通常時の描画（Phase 3: 優先度 攻撃 > 被弾 > 移動 > アイドルブリーズ）
-    const isBlinkOff = player.isInvincible && Math.floor(now / 100) % 2 === 1;
-
-    // パワーオーラ描画（スプライトの背面）
-    if (!isBlinkOff) {
-      drawPlayerAura(ctx, playerScreen.x, playerScreen.y, viewport.tileSize, player.level, player.playerClass, now);
-
-      // シールド輝き描画（maxHp強化時）
-      if (rewardEffects.hasShieldGlow) {
-        drawShieldGlow(ctx, playerScreen.x, playerScreen.y, viewport.tileSize, now);
-      }
-
-      // 残像描画（移動速度強化時、スプライト背面に描画）
-      if (rewardEffects.hasAfterImage) {
-        for (const img of afterImageManagerRef.current.getAfterImages()) {
-          const imgScreen = toScreenPosition(img);
-          drawAfterImage(ctx, imgScreen.x, imgScreen.y, viewport.tileSize, img.alpha);
-        }
-      }
-    }
-
-    if (!isBlinkOff) {
-      const pClass = player.playerClass as 'warrior' | 'thief';
-      const pDir = player.direction as 'down' | 'up' | 'left' | 'right';
-      const playerDrawX = playerScreen.x - playerDrawSize / 2;
-      const playerDrawY = playerScreen.y - playerDrawSize / 2;
-
-      const isAttacking = now < playerAttackUntilRef.current;
-      const isDamaged = now < playerDamageUntilRef.current;
-      const isMoving = movementStateRef.current.activeDirection !== null;
-
-      if (isAttacking) {
-        // 攻撃アニメーション
-        const attackSheets = pClass === 'warrior' ? WARRIOR_ATTACK_SPRITE_SHEETS : THIEF_ATTACK_SPRITE_SHEETS;
-        const attackSheet = attackSheets[pDir];
-        const attackFrameIndex = Math.floor(now / attackSheet.frameDuration) % attackSheet.sprites.length;
-        spriteRenderer.drawSprite(ctx, attackSheet.sprites[attackFrameIndex], playerDrawX, playerDrawY, spriteScale);
-
-        // 武器光跡描画（攻撃アニメーション中のみ）
-        const attackDuration = playerAttackUntilRef.current - (playerAttackUntilRef.current - 300);
-        const attackElapsed = now - (playerAttackUntilRef.current - 300);
-        const attackProgress = Math.min(1, Math.max(0, attackElapsed / attackDuration));
-        drawWeaponTrail(ctx, playerScreen.x, playerScreen.y, viewport.tileSize, player.direction, player.stats.attackPower, player.playerClass, attackProgress);
-
-        // 衝撃波描画（RADIANT ティアのみ、攻撃ヒット時）
-        if (getWeaponTier(player.stats.attackPower) === WeaponTier.RADIANT) {
-          drawShockwave(ctx, playerScreen.x, playerScreen.y, viewport.tileSize, attackElapsed);
-        }
-      } else if (isDamaged) {
-        // 被弾フレーム（200ms表示）
-        const damageSprites = pClass === 'warrior' ? WARRIOR_DAMAGE_SPRITES : THIEF_DAMAGE_SPRITES;
-        spriteRenderer.drawSprite(ctx, damageSprites[pDir], playerDrawX, playerDrawY, spriteScale);
-      } else if (isMoving) {
-        // 歩行アニメーション
-        const playerSheet = getPlayerSpriteSheet(pClass, pDir);
-        const walkFrameIndex = Math.floor(now / playerSheet.frameDuration) % 2;
-        spriteRenderer.drawSprite(ctx, playerSheet.sprites[1 + walkFrameIndex], playerDrawX, playerDrawY, spriteScale);
-
-        // 残像記録（移動速度強化時）
-        if (rewardEffects.hasAfterImage) {
-          afterImageManagerRef.current.recordPosition(player.x, player.y, player.direction, now);
-        }
-      } else {
-        // アイドルブリーズアニメーション
-        const idleSheets = pClass === 'warrior' ? WARRIOR_IDLE_SPRITE_SHEETS : THIEF_IDLE_SPRITE_SHEETS;
-        const idleSheet = idleSheets[pDir];
-        const idleFrameIndex = Math.floor(now / idleSheet.frameDuration) % idleSheet.sprites.length;
-        spriteRenderer.drawSprite(ctx, idleSheet.sprites[idleFrameIndex], playerDrawX, playerDrawY, spriteScale);
-      }
-
-      // 回転パーティクル描画（攻撃速度強化時、常時微小表示）
-      if (rewardEffects.hasSpinParticles) {
-        drawSpinParticles(ctx, playerScreen.x, playerScreen.y, viewport.tileSize, now);
-      }
-
-      // 回復パーティクル描画（回復量強化時）
-      if (rewardEffects.hasHealParticles) {
-        drawHealParticles(ctx, playerScreen.x, playerScreen.y, viewport.tileSize, now);
-      }
-    }
-  }
+  // プレイヤー描画（プレイヤー本体・オーラ・シールド・残像・武器光跡・衝撃波・パーティクル）
+  drawPlayer(frame);
 
   // 低HP警告描画（Phase 4: HP 25%以下でビネットパルス）
   if (player.hp > 0 && player.hp / player.maxHp <= 0.25) {
